# contracts/openapi/mediaflow.v1.yaml
openapi: 3.1.2
info:
  title: MediaFlow API
  version: v1
  description: Contract-first API for MediaFlow identity, inbox scanning, continuous discovery, deterministic identification, review administration, formal catalog reads, downloader management, and source automation.
servers:
  - url: /
paths:
...
  /api/v1/downloader-connections:
    get:
      operationId: listDownloaderConnections
      summary: List redacted downloader connections
      security: [{ cookieAuth: [] }]
      parameters: [{ $ref: '#/components/parameters/Cursor' }, { $ref: '#/components/parameters/Limit' }]
      responses:
        '200': { $ref: '#/components/responses/DownloaderConnectionPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createDownloaderConnection
      summary: Create an encrypted downloader connection
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/DownloaderConnectionInput' }
      responses:
        '201': { $ref: '#/components/responses/DownloaderConnection' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/downloader-connections/connection-tests:
    post:
      operationId: testDownloaderConnection
      summary: Test candidate downloader credentials without saving them
      security: *write-security
      parameters: *write-parameters
      requestBody: { $ref: '#/components/requestBodies/DownloaderConnectionInput' }
      responses:
        '200': { $ref: '#/components/responses/DownloaderConnectionTestResult' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/downloader-connections/{downloaderConnectionId}:
    parameters: [{ $ref: '#/components/parameters/DownloaderConnectionId' }]
    get:
      operationId: getDownloaderConnection
      summary: Read one redacted downloader connection
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/DownloaderConnection' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateDownloaderConnection
      summary: Replace one encrypted downloader connection
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      requestBody: { $ref: '#/components/requestBodies/DownloaderConnectionInput' }
      responses:
        '200': { $ref: '#/components/responses/DownloaderConnection' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteDownloaderConnection
      summary: Delete local downloader credentials without changing remote tasks
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IfMatch' }
      responses:
        '204': { description: Downloader connection deleted locally. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/download-tasks:
    get:
      operationId: listDownloadTasks
      summary: List MediaFlow-owned download tasks
      security: [{ cookieAuth: [] }]
      parameters:
        - { $ref: '#/components/parameters/Cursor' }
        - { $ref: '#/components/parameters/Limit' }
        - { name: connection_id, in: query, schema: { type: string, format: uuid } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/DownloadTaskStatus' } }
        - { name: q, in: query, schema: { type: string, minLength: 1, maxLength: 200 } }
      responses:
        '200': { $ref: '#/components/responses/DownloadTaskPage' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createDownloadTask
      summary: Accept one replay-safe manual download task
      security: *write-security
      parameters:
        - { $ref: '#/components/parameters/CsrfToken' }
        - { $ref: '#/components/parameters/Origin' }
        - { $ref: '#/components/parameters/SecFetchSite' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody: { $ref: '#/components/requestBodies/CreateDownloadTaskRequest' }
      responses:
        '202': { $ref: '#/components/responses/DownloadTask' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationFailed' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/download-tasks/{downloadTaskId}:
    parameters: [{ $ref: '#/components/parameters/DownloadTaskId' }]
    get:
      operationId: getDownloadTask
      summary: Read one MediaFlow-owned download task
      security: [{ cookieAuth: [] }]
      responses:
        '200': { $ref: '#/components/responses/DownloadTask' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
        '500': { $ref: '#/components/responses/InternalError' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: __Host-mediaflow_session }
    csrfHeader: { type: apiKey, in: header, name: X-CSRF-Token }
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string, minLength: 1, maxLength: 512 } }
    Limit: { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 200, default: 50 } }
...
    DownloaderConnectionId: { name: downloaderConnectionId, in: path, required: true, schema: { type: string, format: uuid } }
    DownloadTaskId: { name: downloadTaskId, in: path, required: true, schema: { type: string, format: uuid } }
...
    CsrfToken: { name: X-CSRF-Token, in: header, required: true, schema: { type: string, minLength: 1 } }
    Origin: { name: Origin, in: header, required: false, description: Exact public origin proof used when Fetch Metadata is absent and checked whenever supplied., schema: { type: string, format: uri } }
    SecFetchSite: { name: Sec-Fetch-Site, in: header, required: false, description: Browser Fetch Metadata primary same-origin proof; cross-site and malformed values fail closed., schema: { type: string, enum: [same-origin, same-site, cross-site, none] } }
    IdempotencyKey: { name: Idempotency-Key, in: header, required: true, schema: { type: string, minLength: 1, maxLength: 255 } }
    IfMatch: { name: If-Match, in: header, required: true, description: Decimal configuration version expected by the caller., schema: { type: string, pattern: '^[0-9]{1,19}$' } }
...
  requestBodies:
...
    DownloaderConnectionInput: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/DownloaderConnectionInput' } } } }
    CreateDownloadTaskRequest: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/CreateDownloadTaskRequest' } } } }
...
  responses:
...
    DownloaderConnectionPage: { description: Redacted downloader connections, content: { application/json: { schema: { $ref: '#/components/schemas/DownloaderConnectionPage' } } } }
    DownloaderConnection: { description: One redacted downloader connection, content: { application/json: { schema: { $ref: '#/components/schemas/DownloaderConnection' } } } }
    DownloaderConnectionTestResult: { description: Non-persistent downloader connection test result, content: { application/json: { schema: { $ref: '#/components/schemas/DownloaderConnectionTestResult' } } } }
    DownloadTaskPage: { description: MediaFlow-owned download tasks, content: { application/json: { schema: { $ref: '#/components/schemas/DownloadTaskPage' } } } }
    DownloadTask: { description: One redacted MediaFlow-owned download task, content: { application/json: { schema: { $ref: '#/components/schemas/DownloadTask' } } } }
...
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Forbidden: { description: CSRF or Origin rejected, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    Conflict: { description: Request conflicts with current state, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ValidationFailed: { description: Request validation failed, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    RateLimited: { description: Request rate limited, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    ServiceUnavailable: { description: Service is not ready, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
    InternalError: { description: Internal error, content: { application/json: { schema: { $ref: '#/components/schemas/ApiErrorBody' } } } }
  schemas:
    ApiErrorBody:
      description: Allow-listed API failure envelope with a correlation request ID and optional safe structured details.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, request_id]
          properties:
            code: { type: string, enum: [bootstrap.already_completed, bootstrap.invalid_secret, auth.invalid_credentials, auth.rate_limited, session.expired, csrf.invalid, origin.untrusted, root.not_found, root.unavailable, path.invalid, path.escape, path.symlink_forbidden, inbox.overlap, inbox.not_found, task.not_found, task.invalid_state, task.lease_lost, integration.not_configured, integration.unauthorized, integration.rate_limited, integration.unsupported-version, provider.unavailable, provider.timeout, provider.response-too-large, provider.invalid-response, download.correlation-ambiguous, download.remote-missing, automation.source-disabled, automation.signature-invalid, automation.replay, automation.action-invalid, automation.payload-invalid, automation.downstream-conflict, organization.root-read-only, organization.target-overlap, organization.target-unavailable, organization.path-outside-root, organization.target-exists, organization.source-changed, organization.hardlink-cross-device, organization.operation-unsupported, organization.journal-ambiguous, revision.changed, event.cursor_expired, request.conflict, validation.failed, internal.error] }
            message: { type: string }
            request_id: { type: string, format: uuid }
            details: { type: object, additionalProperties: { type: [string, number, boolean, 'null'] } }
...
    IntegrationHealth:
      description: Stable redacted health classification shared by built-in integrations.
      type: string
      enum: [unconfigured, healthy, degraded, unavailable, unauthorized, rate-limited]
...
    DownloaderKind:
      description: Built-in downloader product selected for one connection.
      type: string
      enum: [qbittorrent, transmission]
    DownloaderFailureCode:
      description: Stable redacted failure classification for downloader connections and MediaFlow-owned download tasks.
      type: string
      enum: [integration.not-configured, integration.unauthorized, integration.rate-limited, integration.unavailable, integration.unsupported-version, provider.timeout, provider.response-too-large, provider.invalid-response, download.correlation-ambiguous, download.remote-missing]
    DownloaderCapabilities:
      description: Negotiated product and protocol capabilities required by the downloader management slice.
      type: object
      additionalProperties: false
      required: [manual_add, task_monitoring, product_version, api_version]
      properties:
        manual_add: { type: boolean }
        task_monitoring: { type: boolean }
        product_version: { type: string, minLength: 1, maxLength: 64 }
        api_version: { type: string, minLength: 1, maxLength: 64 }
    DownloaderConnectionInput:
      description: Candidate downloader configuration whose credentials are write-only and never returned.
      type: object
      additionalProperties: false
      required: [kind, display_name, base_url, username, password, enabled]
      properties:
        kind: { $ref: '#/components/schemas/DownloaderKind' }
        display_name: { type: string, minLength: 1, maxLength: 120 }
        base_url: { type: string, format: uri, minLength: 1, maxLength: 2048 }
        username: { type: string, maxLength: 256, writeOnly: true }
        password: { type: string, maxLength: 4096, writeOnly: true }
        enabled: { type: boolean }
    DownloaderConnection:
      description: Redacted versioned downloader connection, capability, and health projection.
      type: object
      additionalProperties: false
      required: [id, kind, display_name, base_url, enabled, config_version, capabilities, health, checked_at, failure_code, updated_at]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/DownloaderKind' }
        display_name: { type: string, minLength: 1, maxLength: 120 }
        base_url: { type: string, format: uri, maxLength: 2048 }
        enabled: { type: boolean }
        config_version: { type: integer, minimum: 1, maximum: 9007199254740991 }
        capabilities: { oneOf: [{ $ref: '#/components/schemas/DownloaderCapabilities' }, { type: 'null' }] }
        health: { $ref: '#/components/schemas/IntegrationHealth' }
        checked_at: { type: [string, 'null'], format: date-time }
        failure_code: { oneOf: [{ $ref: '#/components/schemas/DownloaderFailureCode' }, { type: 'null' }] }
        updated_at: { type: string, format: date-time }
    DownloaderConnectionPage:
      description: Stable cursor page of redacted downloader connections.
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/DownloaderConnection' } }
        next_cursor: { type: [string, 'null'], maxLength: 512 }
    DownloaderConnectionTestResult:
      description: Non-persistent product, protocol, capability, and redacted health result for candidate credentials.
      type: object
      additionalProperties: false
      required: [reachable, health, capabilities, failure_code, checked_at]
      properties:
        reachable: { type: boolean }
        health: { $ref: '#/components/schemas/IntegrationHealth' }
        capabilities: { oneOf: [{ $ref: '#/components/schemas/DownloaderCapabilities' }, { type: 'null' }] }
        failure_code: { oneOf: [{ $ref: '#/components/schemas/DownloaderFailureCode' }, { type: 'null' }] }
        checked_at: { type: string, format: date-time }
    DownloadTaskStatus:
      description: Durable local submission and monitoring lifecycle.
      type: string
      enum: [queued, submitting, monitoring, retry-wait, completed, failed]
    RemoteDownloadStatus:
      description: Product-neutral remote torrent state.
      type: string
      enum: [queued, downloading, paused, completed, failed, unknown]
    CreateDownloadTaskRequest:
      description: One encrypted manual magnet or HTTPS torrent source associated with a downloader connection.
      type: object
      additionalProperties: false
      required: [connection_id, source, display_name]
      properties:
        connection_id: { type: string, format: uuid }
        source: { type: string, format: uri, minLength: 1, maxLength: 8192, writeOnly: true }
        display_name: { type: string, minLength: 1, maxLength: 512 }
    DownloadTask:
      description: Redacted MediaFlow-owned download task without source, tracker, credential, or remote path fields.
      type: object
      additionalProperties: false
      required: [id, connection_id, connection_display_name, display_name, status, remote_status, progress_basis_points, failure_code, retry_at, linked, projection_version, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        connection_id: { type: string, format: uuid }
        connection_display_name: { type: string, minLength: 1, maxLength: 120 }
        display_name: { type: string, minLength: 1, maxLength: 512 }
        status: { $ref: '#/components/schemas/DownloadTaskStatus' }
        remote_status: { oneOf: [{ $ref: '#/components/schemas/RemoteDownloadStatus' }, { type: 'null' }] }
        progress_basis_points: { type: integer, minimum: 0, maximum: 10000 }
        failure_code: { oneOf: [{ $ref: '#/components/schemas/DownloaderFailureCode' }, { type: 'null' }] }
        retry_at: { type: [string, 'null'], format: date-time }
        linked: { type: boolean }
        projection_version: { type: integer, minimum: 1, maximum: 9007199254740991 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    DownloadTaskPage:
      description: Stable cursor page of MediaFlow-owned download tasks.
      type: object
      additionalProperties: false
      required: [items, next_cursor]
      properties:
        items: { type: array, maxItems: 200, items: { $ref: '#/components/schemas/DownloadTask' } }
        next_cursor: { type: [string, 'null'], maxLength: 512 }
...